requirements.txt
fastapi
uvicorn[standard]
redis
openai
python-dotenv
slowapi
google-cloud-logging
dotenv
flask
